+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - [2026-05-16 22:33] home/godmode/config/env.sh
 - [2026-05-16 22:36] data/data/com.termux/files/home/godmode/lib/encoder.py
 - [2026-05-16 22:37] home/godmode/lib/parsers.py
+- [2026-05-16 22:38] home/godmode/engine/racer.py